Add Feature

+++ WareHouseManagement/Feature/Warehouses/RemoveWarehouse.cs
@@ -15,15 +15,19 @@ namespace WareHouseManagement.Feature.Warehouses {
         }
         [Authorize(Roles = Permission.Admin + "," + Permission.Warehouse)]
         private static async Task<IResult> Handler([FromBody] Request request, ApplicationDbContext context, ClaimsPrincipal user) {
-            var service = context.Users
+            var serviceId = context.Users
                 .Include(u => u.ServiceRegistered)
                 .Where(u => u.UserName == user.Identity.Name)
-                .Select(u => u.ServiceRegistered)
+                .Select(u => u.ServiceId)
                 .FirstOrDefault();
             var warehouse = await context.Warehouses
-                .Where(u => u.ServiceRegisteredFrom.Id == service.Id)
+                .Include(w=>w.Stocks)
+                .Where(u => u.ServiceId == serviceId)
                 .FirstOrDefaultAsync(u => u.Id == request.id);
             if (warehouse != null) {
+                if (warehouse.Stocks.Count !=0) {
+                    return Results.BadRequest(new Response(false, "Lỗi đã xảy ra!"));
+                }
                 context.Warehouses.Remove(warehouse);
                 var result = await context.SaveChangesAsync();
                 if (result > 0)
